pyproject.toml
[tool.poetry]
name = "audio-scraper"
version = "0.1.0"
description = "Scrapes and processes speech audio from web sources."
authors = ["Stephen Melsom <sjmelsom@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "audio_scraper"}]

[tool.poetry.dependencies]
python = "^3.10"
yt-dlp = "^2022.11.11"
pydub = "^0.25.1"
toml = "^0.10.2"
loguru = "^0.6.0"
setuptools-rust = "^1.5.2"
openai-whisper = "^20230117"
tqdm = "^4.64.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pendulum = "^2.1.2"
pytest = "^7.2.0"
pylint = "^2.15.9"
isort = "^5.11.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
scrape = "audio_scraper.pipeline:run_all"
refilter = "audio_scraper.pipeline:refilter"
transcribe = "audio_scraper.transcriber:transcribe"
